+++ flightplan.js
@@ -103,7 +103,6 @@ plan.local(['e2e', 'unit'], function(local) {
     // Install openveo-core dependencies and compile sources
     local.with('cd ' + workingDirectory, function() {
       local.exec('npm install --ignore-scripts');
-      local.exec('bower install');
       local.exec('grunt prod');
     });